requirements.txt
xpander-sdk
xpander-utils
boto3
python-dotenv